requirements.txt
pandas
numpy
streamlit
scikit-learn
sentence-transformers
google-generativeai>=0.5.0
python-dotenv
